Distinguish between old and new character controllers
