+++ Assets/Code/Entities/Penguin/PenguinBlob.cs
@@ -44,7 +44,7 @@ namespace PQ.Entities.Penguin
         [Header("Component References")]
         [SerializeField] private Rigidbody2D _penguinRigidbody;
         [SerializeField] private PenguinAnimation _penguinAnimation;
-        [SerializeField] private CharacterController2D _characterController;
+        [SerializeField] private CharacterController2D_old _characterController;
 
         [Header("Collider References")]
         [SerializeField] private BoxCollider2D     _boundingBoxCollider;
@@ -60,7 +60,7 @@ namespace PQ.Entities.Penguin
         public CharacterController2DSettings OnBellySettings => _penguinOnFeetSettings;
 
         public PenguinAnimation      Animation            => _penguinAnimation;
-        public CharacterController2D CharacterController  => _characterController;
+        public CharacterController2D_old CharacterController  => _characterController;
         public Rigidbody2D           Rigidbody            => _penguinRigidbody;
         public Vector2               SkeletalRootPosition => _penguinAnimation.SkeletalRootPosition;
         public Vector2               CenterOfMass         => _penguinRigidbody.worldCenterOfMass;
