--- Assets/Code/_Common/Physics.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e1ccd3978ef3d4047989ac1eb96b2996
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
